Apply pdfium_noshorten_config to xfa/fxfa/parser

+++ xfa/fxfa/parser/cxfa_box.h
@@ -27,7 +27,7 @@ class CXFA_Box : public CXFA_Node {
   XFA_AttributeValue GetPresence();
   std::tuple<XFA_AttributeValue, bool, float> Get3DStyle();
 
-  int32_t CountEdges();
+  size_t CountEdges();
   CXFA_Edge* GetEdgeIfExists(int32_t nIndex);
   CXFA_Fill* GetOrCreateFillIfPossible();
